+++ packages/toolbar/CHANGELOG.md
@@ -1,5 +1,28 @@
 # Changelog
 
+## [2.5.3-beta.1](https://github.com/launchdarkly/launchdarkly-toolbar/compare/2.5.2-beta.1...2.5.3-beta.1) (2025-12-17)
+
+
+### Bug Fixes
+
+* [REL-11336] Update inputs to stop propagation ([#441](https://github.com/launchdarkly/launchdarkly-toolbar/issues/441)) ([d0963e4](https://github.com/launchdarkly/launchdarkly-toolbar/commit/d0963e4715d4cdd464ed59519b14ff1b9fa725d2))
+
+## [2.5.2-beta.1](https://github.com/launchdarkly/launchdarkly-toolbar/compare/2.5.1-beta.1...2.5.2-beta.1) (2025-12-16)
+
+
+### Bug Fixes
+
+* [REL-11306] handle host application hot reloads better ([#438](https://github.com/launchdarkly/launchdarkly-toolbar/issues/438)) ([e825036](https://github.com/launchdarkly/launchdarkly-toolbar/commit/e8250367c2381030199fa6fb1aff736ad469f8f3))
+* handle host application hot reloads better ([e825036](https://github.com/launchdarkly/launchdarkly-toolbar/commit/e8250367c2381030199fa6fb1aff736ad469f8f3))
+
+## [2.5.1-beta.1](https://github.com/launchdarkly/launchdarkly-toolbar/compare/2.5.0-beta.1...2.5.1-beta.1) (2025-12-16)
+
+
+### Bug Fixes
+
+* [REL-11305] wait for context to update to load toolbar ([#436](https://github.com/launchdarkly/launchdarkly-toolbar/issues/436)) ([46fbbe1](https://github.com/launchdarkly/launchdarkly-toolbar/commit/46fbbe153450c09f67e05019e1f704031535dcf3))
+* wait for context to update to load toolbar ([46fbbe1](https://github.com/launchdarkly/launchdarkly-toolbar/commit/46fbbe153450c09f67e05019e1f704031535dcf3))
+
 ## [2.5.0-beta.1](https://github.com/launchdarkly/launchdarkly-toolbar/compare/2.4.1-beta.1...2.5.0-beta.1) (2025-12-15)